Clean up __init__
