+++ src/quacc/utils/dicts.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 from copy import deepcopy
 from typing import TYPE_CHECKING
 
-from quacc import Remove
-
 if TYPE_CHECKING:
     from typing import Any
 
@@ -161,3 +159,20 @@ def clean_task_doc(start_dict: dict[str, Any]) -> dict[str, Any]:
         Cleaned dictionary
     """
     return sort_dict(remove_dict_entries(start_dict, None))
+
+
+class Remove:
+    """
+    A sentinel class used in quacc to mark a key in a dictionary for removal.
+
+    Note: This is more robust than using `None` as the sentinel value because
+    `None` is a valid value for many keyword arguments. Also, using `object()`
+    as the sentinel value is not robust because its value changes every time
+    it is instantiated, which means an `object()` provided by the user locally
+    will not match an `object()` instantiated on the remote machine.
+    """
+
+    def __init__(self):
+        raise NotImplementedError(
+            "Remove is a sentinel class and should not be instantiated."
+        )
